+++ src/briefy/ws/utils/user.py
@@ -1,9 +1,11 @@
 """User utilities for briefy.webservice."""
 from briefy.common.utils.cache import timeout_cache
+from briefy.ws import logger
 from briefy.ws.config import USER_SERVICE_BASE
 from briefy.ws.config import USER_SERVICE_TIMEOUT
 
 import requests
+import transaction
 
 
 def _get_user_info_from_service(user_id: str) -> dict:
@@ -19,13 +21,22 @@ def _get_user_info_from_service(user_id: str) -> dict:
     )
     # TODO: improve this to user current user locale
     headers = {'X-Locale': 'en_GB'}
-    resp = requests.get(
-        endpoint,
-        headers=headers
-    )
-    if resp.status_code == 200:
-        raw_data = resp.json()
-        data = raw_data['data'] if 'data' in raw_data else data
+    savepoint = transaction.savepoint()
+    try:
+        resp = requests.get(
+            endpoint,
+            headers=headers
+        )
+    except ConnectionError as exc:
+        logger.warn('Failure connecting to internal user service. Exception: {exc}'.format(exc=exc))
+        savepoint.rollback()
+    else:
+        if resp.status_code == 200:
+            raw_data = resp.json()
+            data = raw_data['data'] if 'data' in raw_data else data
+        else:
+            msg = 'Getting user info from internal services fail. Status code: {status_code}.'
+            logger.info(msg.format(status_code=resp.status_code))
     return data